currencyproject_backend/requirements.txt
numpy
scikit-learn
MetaTrader5
django
djangorestframework
numpy<2
django-cors-headers
